WebcamCapture now properly reduxed
