+++ src/components/App/actionTypes.js
@@ -1,3 +1,2 @@
 export const SET_USERS = 'SET_USERS';
 export const SET_SLACK_USER_FETCH_ERROR = 'SET_SLACK_USER_FETCH_ERROR';
-export const SET_PREDICTION = 'SET_PREDICTION';
