+++ src/components/App/actions.js
@@ -1,8 +1,4 @@
-import {
-	SET_USERS,
-	SET_SLACK_USER_FETCH_ERROR,
-	SET_PREDICTION
-} from './actionTypes';
+import {SET_USERS, SET_SLACK_USER_FETCH_ERROR} from './actionTypes';
 
 import {loadUsers as reallyLoadUsers} from './../../utils/slack';
 
@@ -17,13 +13,6 @@ export function setSlackUserFetchError(slackUserFetchError) {
 	return {
 		type: SET_SLACK_USER_FETCH_ERROR,
 		slackUserFetchError
-	};
-}
-
-export function setPrediction(prediction) {
-	return {
-		type: SET_PREDICTION,
-		prediction
 	};
 }
