+++ src/components/WebcamCapture/actions.js
@@ -1,0 +1,8 @@
+import {SET_PREDICTION} from './actionTypes';
+
+export function setPrediction(prediction) {
+	return {
+		type: SET_PREDICTION,
+		prediction
+	};
+}
