+++ src/components/WebcamCapture/reducer.js
@@ -1,0 +1,6 @@
+import {SET_PREDICTION} from './actionTypes';
+
+export function prediction(state = null, action) {
+	if (action.type === SET_PREDICTION) return action.prediction;
+	return state;
+}
